// src/app/ProjectFiles/Models/NominatimAddress.ts
export interface NominatimAddress {
    administrative: string;
    city: string;
    town: string;
    village: string;
    hamlet: string;
    country: string;
    house_number: string;
    country_code: string;
    postcode: string;
    road: string;
    state: string;
    waterway: string;
    railway: string;
}